refactor: use relative root compat paths
