+++ rustfs/src/runtime_capabilities.rs
@@ -18,7 +18,7 @@ use rustfs_storage_api::{
     UserspaceProfilingCapability,
 };
 
-use crate::runtime_capabilities_storage_compat::{EndpointServerPools, topology_snapshot_from_endpoint_pools_with_capabilities};
+use super::runtime_capabilities_storage_compat::{EndpointServerPools, topology_snapshot_from_endpoint_pools_with_capabilities};
 
 const NOT_WIRED_INTO_RUNTIME: &str = "not wired into runtime";
 const STORAGE_MEDIA_NOT_REPORTED: &str = "storage media not reported by endpoints";
@@ -131,8 +131,8 @@ fn cgroup_memory_status() -> CapabilityStatus {
 
 #[cfg(test)]
 mod tests {
+    use super::super::runtime_capabilities_storage_compat::{Endpoint, Endpoints, PoolEndpoints};
     use super::*;
-    use crate::runtime_capabilities_storage_compat::{Endpoint, Endpoints, PoolEndpoints};
     use rustfs_storage_api::{CapabilityState, ObservabilitySnapshotProvider, TopologySnapshotProvider};
 
     #[tokio::test]
