+++ rustfs/src/storage/mod.rs
@@ -30,11 +30,11 @@ pub mod timeout_wrapper;
 pub mod tonic_service;
 
 pub(crate) type StorageDeletedObject = rustfs_storage_api::DeletedObject;
-pub(crate) type StorageGetObjectReader = crate::storage::core_storage_compat::GetObjectReader;
-pub(crate) type StorageObjectInfo = crate::storage::core_storage_compat::ObjectInfo;
-pub(crate) type StorageObjectOptions = crate::storage::core_storage_compat::ObjectOptions;
+pub(crate) type StorageGetObjectReader = self::core_storage_compat::GetObjectReader;
+pub(crate) type StorageObjectInfo = self::core_storage_compat::ObjectInfo;
+pub(crate) type StorageObjectOptions = self::core_storage_compat::ObjectOptions;
 pub(crate) type StorageObjectToDelete = rustfs_storage_api::ObjectToDelete;
-pub(crate) type StoragePutObjReader = crate::storage::core_storage_compat::PutObjReader;
+pub(crate) type StoragePutObjReader = self::core_storage_compat::PutObjReader;
 
 #[cfg(test)]
 mod concurrent_fix_test;
