+++ rustfs/src/workload_admission.rs
@@ -17,10 +17,10 @@ use rustfs_concurrency::{
     WorkloadClass,
 };
 
-use crate::storage::concurrency::get_concurrency_manager;
-use crate::workload_admission_storage_compat::{
+use super::workload_admission_storage_compat::{
     get_global_bucket_metadata_sys, get_global_replication_pool, replication_queue_current_count,
 };
+use crate::storage::concurrency::get_concurrency_manager;
 
 const BUCKET_METADATA_RUNTIME_NOT_INITIALIZED: &str = "bucket metadata runtime not initialized";
 const HEAL_MANAGER_NOT_INITIALIZED: &str = "heal manager not initialized";
